pyproject.toml
[project]
name = "APE"
description = "LLM Deception and Persuasion"
requires-python = ">=3.11"
version = "0.0.1"

dependencies = [
    "python-dotenv",
    "openai",
    "matplotlib",
    "pytz",
    "pyyaml",
    "scipy",
    "pandas",
    "litellm",
    "google-cloud-aiplatform",
    "seaborn",
    "scikit-learn",
    "transformers",
    "torch",
    "accelerate",
    "tabulate",
    "hydra-core"
]

[project.optional-dependencies]
dev = [
    "black",
    "pre-commit",
    "mypy",
    "isort",
    "pandas-stubs",
    "types-tabulate",
    "types-tqdm"
]
test = [
    "pytest",
    "pytest-asyncio"
]

[tool.logviz-dependencies]
Flask = "^2.3.2"
jsonlines = "^3.1.0"
graphene = "^3.3"
graphql-server = { version = "^3.0.0b7", extras = ["flask"] }
flask-cors = "^4.0.0"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["persuasion*"]

[tool.setuptools.package-data]
"*" = ["*.jsonl", "*.txt", "*.csv"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = [
    "src"
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
ignore_missing_imports = true

# Exclude files with many errors for now
exclude = [
    "src/visualizations/.*",
    "src/test_visualization_recreation.py"
]
